feat(mssql): Format storage into domains and separate these into different files
