+++ .changeset/mighty-mssql-features.md
@@ -2,7 +2,7 @@
 "@mastra/mssql": minor
 ---
 
-feat(mssql): implement comprehensive scoring system and enhance storage operations
+feat(mssql): implement comprehensive scoring system, enhance storage operations & format storage into domains
 
 - Add full CRUD operations for scoring system (getScoreById, saveScore, getScoresByScorerId, etc.)
 - Implement message deletion functionality with thread timestamp updates
@@ -14,5 +14,6 @@ feat(mssql): implement comprehensive scoring system and enhance storage operatio
 - Add float data type support in schema creation
 - Enhance timestamp handling with proper SQL DateTime2 usage
 - Update core dependency to latest version
+- Format storage into domains and separate these into different files
 
 BREAKING CHANGE: deleteMessages support is now enabled (was previously false)
